Refactor RequestEnvironment
RequestEnvironmet now allows to be updated with additional variables
that are not pulled directly from the environment.

Since the type now allows mutation, the instance method now returns a
copy instead of a refernce to prevent mutation of global state. At the
same time 'instance' is also depreacted in favour of 'make' to highlight
that this is not our usual singelton pattern but just a construction
that caches some "expensive" to initialize values.